0.4

+++ src/scope/ArrayClassScope.cpp
@@ -5,7 +5,7 @@ PrimitiveClassScope<std::vector<SharedIValue>>(Type::ARRAY_NAME){
     (*getPublicVariables())[*CAPACITY_NAME]=
         std::make_shared<Variable>(
             CAPACITY_NAME,
-            Type::INT,
+            Type::ULONG,
             std::make_shared<bool>(true)
         );
 }
